Updates to multiple classes posts
The previous improvements to the encounter engine allowed reactions for spells like counterspell and shield to be reintroduced into the sorcerer and wizard player characters.

The adventuring day XP data has been updated to reflect these changes, and new figures have been generated for xp-and-player-characters, daily-xp-and-encounter-difficulty, and short-adventuring-days.

The only post that needed any significant changes to the post itself was short-adventuring-days.